+++ src/hotspot/cpu/aarch64/frame_aarch64.inline.hpp
@@ -233,7 +233,7 @@ inline intptr_t* frame::real_fp() const {
 
 inline int frame::frame_size() const {
   return is_interpreted_frame()
-    ? sender_sp() - sp()
+    ? narrow_cast<int>(sender_sp() - sp())
     : cb()->frame_size();
 }
