+++ src/hotspot/share/code/codeBlob.hpp
@@ -137,8 +137,8 @@ public:
   CodeBlob() : _type(compiler_none) {}
 
   // Returns the space needed for CodeBlob
-  static unsigned int allocation_size(CodeBuffer* cb, int header_size);
-  static unsigned int align_code_offset(int offset);
+  static int allocation_size(CodeBuffer* cb, int header_size);
+  static int align_code_offset(int offset);
 
   // Deletion
   virtual void flush();
@@ -192,9 +192,15 @@ public:
   // Sizes
   int size() const                               { return _size; }
   int header_size() const                        { return _header_size; }
-  int relocation_size() const                    { return (address) relocation_end() - (address) relocation_begin(); }
-  int content_size() const                       { return           content_end()    -           content_begin();    }
-  int code_size() const                          { return           code_end()       -           code_begin();       }
+  int relocation_size() const {
+    return narrow_cast<int>((address) relocation_end() - (address) relocation_begin());
+  }
+  int content_size() const {
+    return narrow_cast<int>(content_end() - content_begin());
+  }
+  int code_size() const {
+    return narrow_cast<int>(code_end() - code_begin());
+  }
   // Only used from CodeCache::free_unused_tail() after the Interpreter blob was trimmed
   void adjust_size(size_t used) {
     _size = (int)used;
