+++ src/hotspot/share/gc/shared/markBitMap.inline.hpp
@@ -37,7 +37,8 @@ inline HeapWord* MarkBitMap::get_next_marked_addr(const HeapWord* const addr,
                                                   HeapWord* const limit) const {
   assert(limit != NULL, "limit must not be NULL");
   // Round addr up to a possible object boundary to be safe.
-  size_t const addr_offset = addr_to_offset(align_up(addr, HeapWordSize << _shifter));
+  size_t const addr_offset = addr_to_offset(
+      align_up(addr, static_cast<uint>(HeapWordSize << _shifter)));
   size_t const limit_offset = addr_to_offset(limit);
   size_t const nextOffset = _bm.get_next_one_offset(addr_offset, limit_offset);
   return offset_to_addr(nextOffset);
