+++ src/hotspot/share/oops/oop.inline.hpp
@@ -157,7 +157,7 @@ size_t oopDesc::size_given_klass(Klass* klass)  {
 
   if (lh > Klass::_lh_neutral_value) {
     if (!Klass::layout_helper_needs_slow_path(lh)) {
-      s = lh >> LogHeapWordSize;  // deliver size scaled by wordSize
+      s = static_cast<size_t>(lh >> LogHeapWordSize);  // deliver size scaled by wordSize
     } else {
       s = klass->oop_size(this);
     }
@@ -170,12 +170,12 @@ size_t oopDesc::size_given_klass(Klass* klass)  {
       size_t size_in_bytes;
       size_t array_length = (size_t) ((arrayOop)this)->length();
       size_in_bytes = array_length << Klass::layout_helper_log2_element_size(lh);
-      size_in_bytes += Klass::layout_helper_header_size(lh);
+      size_in_bytes += static_cast<size_t>(Klass::layout_helper_header_size(lh));
 
       // This code could be simplified, but by keeping array_header_in_bytes
       // in units of bytes and doing it this way we can round up just once,
       // skipping the intermediate round to HeapWordSize.
-      s = align_up(size_in_bytes, MinObjAlignmentInBytes) / HeapWordSize;
+      s = align_up(size_in_bytes, static_cast<uint>(MinObjAlignmentInBytes)) / HeapWordSize;
 
       assert(s == klass->oop_size(this) || size_might_change(), "wrong array object size");
     } else {
