+++ src/hotspot/share/oops/typeArrayOop.hpp
@@ -124,7 +124,7 @@ private:
 
     julong size_in_bytes = (juint)length;
     size_in_bytes <<= element_shift;
-    size_in_bytes += instance_header_size;
+    size_in_bytes += static_cast<uint>(instance_header_size);
     julong size_in_words = ((size_in_bytes + (HeapWordSize-1)) >> LogHeapWordSize);
     assert(size_in_words <= (julong)max_jint, "no overflow");
